requirements.txt
nltk==3.9.1
pyTelegramBotAPI
python-dotenv
